Variable_scope compat

+++ src/04_imdb.py
@@ -82,7 +82,7 @@ def model_fn(mode, inputs, reuse=False):
                                            num_cells=[FLAGS.rnn_cells] * FLAGS.rnn_layers,
                                            batch_size=FLAGS.batch_size)
 
-        rnn_outputs, rnn_states = tf.nn.dynamic_rnn(cell, samples, dtype=tf.int32, initial_state=initial_state)
+        rnn_outputs, rnn_states = tf.nn.dynamic_rnn(cell, samples, dtype=tf.float32, initial_state=initial_state)
 
         # Split the outputs of the RNN into the actual outputs and the state update gate
         rnn_outputs, updated_states = split_rnn_outputs(FLAGS.model, rnn_outputs)
